VibeCodingProjects/Templates/Backend/Python-FastAPI/starter/requirements.txt
# ============================================
# Dosya: requirements.txt
# Amaç: Python bağımlılıkları — FastAPI, SQLAlchemy, JWT, CORS
# ============================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# CORS & Multipart
python-multipart==0.0.9

# Testing
pytest==8.0.1
pytest-asyncio==0.23.4
httpx==0.27.0

# Utils
python-dateutil==2.8.2
